+++ database/migrations/2024_01_12_114258_add_client_id_to_internal_agent_my_businesses_table.php
@@ -11,11 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('internal_agent_my_businesses', static function (Blueprint $table) {
-            $table->foreignId('client_id')
-            ->nullable()
-            ->constrained('clients')->after('policy_draft_date');
-        });
+        if (Schema::hasTable('internal_agent_my_businesses')) {
+            Schema::table('internal_agent_my_businesses', static function (Blueprint $table) {
+                if (!Schema::hasColumn('internal_agent_my_businesses', 'client_id')) {
+                    $table->foreignId('client_id')
+                          ->nullable()
+                          ->constrained('clients')
+                          ->after('policy_draft_date');
+                }
+            });
+        }
     }
 
     /**
@@ -23,8 +28,28 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('internal_agent_my_businesses', static function (Blueprint $table) {
-            $table->dropColumn('client_id');
-        });
+        if (Schema::hasTable('internal_agent_my_businesses')) {
+            Schema::table('internal_agent_my_businesses', static function (Blueprint $table) {
+                if (self::hasForeignKey('internal_agent_my_businesses', 'internal_agent_my_businesses_client_id_foreign')) {
+                    $table->dropForeign(['client_id']);
+                }
+                if (Schema::hasColumn('internal_agent_my_businesses', 'client_id')) {
+                    $table->dropColumn('client_id');
+                }
+            });
+        }
+    }
+
+    /**
+     * Check if a foreign key exists on a table.
+     */
+    public static function hasForeignKey($table, $foreignKey): bool
+    {
+        $conn = Schema::getConnection()->getDoctrineSchemaManager();
+        $keys = array_map(function ($key) {
+            return $key->getName();
+        }, $conn->listTableForeignKeys($table));
+
+        return in_array($foreignKey, $keys);
     }
 };
