+++ database/migrations/2024_03_06_115016_add_business_id_in_calls_table.php
@@ -12,10 +12,12 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('calls', static function (Blueprint $table) {
-            $table->foreignId('policy_id')
-                ->after('user_id')
-                ->nullable()
-                ->constrained('calls');
+            if (!Schema::hasColumn('calls', 'policy_id')) { // Check if the column doesn't exist
+                $table->foreignId('policy_id')
+                      ->after('user_id')
+                      ->nullable()
+                      ->constrained('internal_agent_my_businesses'); // Assuming the foreign key is intended to reference the 'policies' table, not 'calls'
+            }
         });
     }
 
@@ -25,8 +27,23 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('calls', static function (Blueprint $table) {
-            $table->dropForeign('policy_id');
-            $table->dropColumn('policy_id');
+            if (Schema::hasColumn('calls', 'policy_id')) { // Check if the column exists
+                if (self::hasForeignKey('calls', 'calls_policy_id_foreign')) {
+                    $table->dropForeign('calls_policy_id_foreign');
+                }
+
+                $table->dropColumn('policy_id');
+            }
         });
+    }
+
+    public static function hasForeignKey($table, $key) {
+        $conn = Schema::getConnection()->getDoctrineSchemaManager();
+
+        $keys =  array_map(function($key) {
+            return $key->getName();
+        }, $conn->listTableForeignKeys($table));
+        
+        return in_array($key, $keys);
     }
 };
